backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
pandas
openpyxl
httpx
python-dotenv
